[#25]♻️ refactor : Exception 구조 변경
