+++ src/main/java/com/sparta/mypet/common/exception/custom/CommentNotFoundException.java
@@ -1,4 +1,4 @@
-package com.sparta.mypet.common.exception;
+package com.sparta.mypet.common.exception.custom;
 
 public class CommentNotFoundException extends RuntimeException {
 	public CommentNotFoundException(String message) {
