+++ src/main/java/com/sparta/mypet/common/exception/custom/PostNotFoundException.java
@@ -1,6 +1,4 @@
-package com.sparta.mypet.common.exception;
-
-import com.sparta.mypet.common.entity.GlobalMessage;
+package com.sparta.mypet.common.exception.custom;
 
 public class PostNotFoundException extends RuntimeException {
 	public PostNotFoundException(String message) {
