+++ src/main/java/com/sparta/mypet/common/exception/custom/UserMisMatchException.java
@@ -1,4 +1,4 @@
-package com.sparta.mypet.common.exception;
+package com.sparta.mypet.common.exception.custom;
 
 public class UserMisMatchException extends RuntimeException {
 	public UserMisMatchException(String message) {
