+++ src/main/java/com/sparta/mypet/domain/like/LikeService.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.sparta.mypet.common.entity.GlobalMessage;
-import com.sparta.mypet.common.exception.PostNotFoundException;
+import com.sparta.mypet.common.exception.custom.PostNotFoundException;
 import com.sparta.mypet.domain.auth.UserRepository;
 import com.sparta.mypet.domain.auth.entity.User;
 import com.sparta.mypet.domain.like.entity.Like;
